rename navbar.js to load_element.js
