+++ srcs/django_code/transcend/myapp/static/js/load_element.js
